Add files via upload
Created a quadrant chart with 2 average lines which is Average Length of stay and average cost of discharge. Then I added a slicer for facility name. Finally, added one percentile learn where the percentile was set to 90%. The screenshot for it shows that only 5 hospitals in New York City has their average cost of discharge higher than $45,365 that does not come in 90% percentile area.