Add files via upload
First dashboard is the length of stay dashboard which comprises of 4 visualizations(Key influencers, line and stacker bar chart, 2 horizontal bar charts) and 4 cards displaying Total surgeons, Average LOS days, Total hospitals and Total discharges. Created multiple filters in each visualization and 1 filter for the dashboard which is Selecting health service area and this filter is synced to all the visualizations of this dashboard.