Add files via upload
This dashboard contains 4 visualizations(Key influencers, Quadrant chart with 3 reference lines and 2 horizontal bar charts showing top 3 and bottom facilities with avg cost per discharge), it has 4 cards defining the average LOS days and average cost per discharge, total hospitals and total discharges. There is 1 filter for the dashboard for selecting a particular area that is synced with all the visualizations of this dashboard.