Add files via upload
This dashboard shows the profile for selected hospital which the user can select from the filter. The profile shows the avg LOS days for the selected hospital compared with the avg LOS of all hospitals combined. Similarly, it shows the avg cost per discharge vs all and 5 more factors that influence the efficiency of hospitals.